tel/requirements.txt
pyrogram==2.0.103
TgCrypto==1.2.5
python-dotenv==1.0.0
requests
loguru